Add basic login and signup page
